### 2018-04-18 #679 @whedon
[PRE REVIEW]: malan: MAle Lineage ANlysis
**Submitting author:** @mikldk (<a href="http://orcid.org/0000-0002-0234-0266">Mikkel Meyer Andersen</a>)
**Repository:** <a href="https://github.com/mikldk/malan" target ="_blank">https://github.com/mikldk/malan</a>
**Version:** v0.0.3
