### 2018-04-18 #679 @mikldk
I see that the footer in the pdf article proof contains the Markdown markup (the stars for making some letters boldface). I am not sure how to fix this -- is there an option to provide a clear text title without markup?
