### 2018-04-18 #679 @arfon
> I see that the footer in the pdf article proof contains the Markdown markup (the stars for making some letters boldface). I am not sure how to fix this -- is there an option to provide a clear text title without markup?

Yeah, sorry about this. We only have limited support for custom typesetting. The only suggestion I have for this is to remove the custom formatting from the title (I realize this isn't ideal).
