### 2018-04-18 #679 @mikldk
@arfon, I guess it would make sense to look into making this possible?

By the look of it, it would require changes to:

[openjournals/whedon/lib/whedon/processor.rb#L101](https://github.com/openjournals/whedon/blob/6d7d3729232081e24ef1f1d03b9203f12ce8771f/lib/whedon/processor.rb#L101)
and
[openjournals/whedon/resources/latex.template#L64](https://github.com/openjournals/whedon/blob/6d7d3729232081e24ef1f1d03b9203f12ce8771f/resources/latex.template#L64)

One way is to add an additional `title_footer` YAML field, but I guess that would have impact many places. The easy fix is to "just" strip the title for Markdown synxtax related to italic/bold. Maybe simply with regex with pairs of `**`/`*`/`__`/`_` like (not tested)

```ruby
gsub(/\*\*(.+?)\*\*/,'\1')
gsub(/\*(.+?)\*/,'\1')
gsub(/__(.+?)__/,'\1')
gsub(/_(.+?)_/,'\1')
```
Or maybe using a proper function like <http://www.rubydoc.info/github/vmg/redcarpet/Redcarpet/Render/StripDown>.
And then in the footer in the LaTeX template have a, `$paper_title_footer$`, say.

Or: If I'm the only one making troubles, I change my title :-).
