### 2018-04-18 #679 @arfon
> @arfon, I guess it would make sense to look into making this possible?

Thanks for digging into this @mikldk - I agree this is how it _could_ work, I'm just not sure this is a direction we want to go in. JOSS is deliberately lightweight in terms of tooling and that extends to how sophisticated we can be with our typesetting.

I suggest you leave your `paper.md` _as is_ and if you remind me when we're accepting the publication I can manually fix this for you when I'm processing the paper.
