### 2018-04-19 #684 @whedon
[REVIEW]: malan: MAle Lineage ANalysis
**Submitting author:** @mikldk (<a href="http://orcid.org/0000-0002-0234-0266">Mikkel Meyer Andersen</a>)
**Repository:** <a href="https://github.com/mikldk/malan" target ="_blank">https://github.com/mikldk/malan</a>
**Version:** v1.0.0
