### 2018-04-19 #679 @arfon
> @arfon, it might be a good general practice to have each paper reviewed by at least two reviewers if possible. This would allow for more scrutiny, without delaying the process too much.

Yes, we do this regularly but leave this up to the discretion of the editor (@brainstorm in this case).
