### 2018-04-22 69c675f
details return if needed
