### 2018-04-23 #684 @mikldk
@brainstorm / @arfon, I had an awfully embarrassing typo with a missing "a" in the title (ANlysis / AN**a**lysis). I have updated my paper.md in my repo and ha @whedon recompile proofs. Can you change the title in the submission (including the title of this issue -> "malan: MAle Lineage ANalysis").
