### 2018-04-23 #684 @arfon
> Can you change the title in the submission (including the title of this issue -> "malan: MAle Lineage ANalysis").

Fixed here. The version on the main JOSS site will be updated when the submission is processed.
