### 2018-04-23 29e9867
Functions to get brothers, uncles and cousins
