+++ man/get_brothers.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/RcppExports.R
+\name{get_brothers}
+\alias{get_brothers}
+\title{Get brothers}
+\usage{
+get_brothers(individual)
+}
+\arguments{
+\item{individual}{individual}
+}
+\value{
+List with brothers
+}
+\description{
+Get individual's brothers
+}
+\seealso{
+\code{\link[=get_uncles]{get_uncles()}}, \code{\link[=get_children]{get_children()}}, \code{\link[=get_cousins]{get_cousins()}}
+}
