+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: malan
 Type: Package
 Title: MAle Lineage ANalysis
-Version: 0.0.3.1
-Date: 2018-04-17
+Version: 0.0.3
+Date: 2018-04-24
 Authors@R: c(
     person(given = "Mikkel Meyer", family = "Andersen", 
         email = "mikl@math.aau.dk", role = c("aut", "cre"))
