+++ .travis.yml
@@ -1,31 +1,51 @@
+# From https://github.com/tidyverse/dplyr/blob/master/.travis.yml
+
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
+# Default configuration for use with tic package
+# Usually you shouldn't need to change the first part of the file
 
-language: R
+# DO NOT CHANGE THE CODE BELOW
+before_install: R -q -e 'install.packages(c("remotes", "curl")); remotes::install_github("ropenscilabs/tic"); tic::prepare_all_stages(); tic::before_install()'
+install: R -q -e 'tic::install()'
+after_install: R -q -e 'tic::after_install()'
+before_script: R -q -e 'tic::before_script()'
+script: R -q -e 'tic::script()'
+after_success: R -q -e 'tic::after_success()'
+after_failure: R -q -e 'tic::after_failure()'
+before_deploy: R -q -e 'tic::before_deploy()'
+deploy:
+  provider: script
+  script: R -q -e 'tic::deploy()'
+  on:
+    all_branches: true
+after_deploy: R -q -e 'tic::after_deploy()'
+after_script: R -q -e 'tic::after_script()'
+# DO NOT CHANGE THE CODE ABOVE
+
+# Custom parts:
+
+# Header
+language: r
 sudo: true
-#cache: packages
+dist: trusty
+cache: packages
 
-r:
-- release
-#- devel
+matrix:
+  include:
+  - r: 3.1
+  - r: 3.2
+  - r: oldrel
+  - r: release
+    env:
+    - BUILD_PKGDOWN: true
+  - r: devel
 
-before_install:
-  # C++14
-  - sudo add-apt-repository -y ppa:ubuntu-toolchain-r/test
-  - sudo apt-get update -qq
-  # R
-  - sudo add-apt-repository -y ppa:marutter/rrutter
-  - sudo apt-get update -qq
+#env
+env:
+  global:
+  - _R_CHECK_FORCE_SUGGESTS_=false
+  - MAKEFLAGS="-j 2"
+  - TRAVIS_CXXFLAGS="-Wall -Wextra -pedantic -Werror"
 
-install: 
-  # C++14
-  - sudo apt-get install -qq g++-5
-  #- sudo update-alternatives --install /usr/bin/g++ g++ /usr/bin/g++-5 90
-  # R
-  - sudo apt-get install r-base r-base-dev gfortran libopenblas-base
-  - sudo apt-get install libudunits2-dev
-  - sudo Rscript -e 'install.packages(c("remotes"), repos = "http://cran.r-project.org")'
-  - sudo Rscript -e 'remotes::install_github("tidyverse/dplyr")'
-  - sudo Rscript -e 'install.packages(c("igraph", "knitr", "rmarkdown", "testthat", "tidygraph", "RcppProgress", "RcppArmadillo", "ggraph", "covr"), repos = "http://cran.r-project.org")'
-
-after_success:
-  - Rscript -e 'library(covr); codecov()'
+#services
+services:
