+++ .travis.yml
@@ -32,8 +32,6 @@ cache: packages
 
 matrix:
   include:
-  - r: 3.1
-  - r: 3.2
   - r: oldrel
   - r: release
     env:
