+++ tic.R
@@ -1,4 +1,3 @@
 # https://github.com/tidyverse/dplyr/blob/master/tic.R
 # Accept warnings from setOldClass() in R 3.1
 add_package_checks(warnings_are_errors = getRversion() > "3.1.3")
-
