### 2018-04-27 93e790e
Estimating theta from subpopulations
