+++ src/api_utility_autosomal.cpp
@@ -465,7 +465,7 @@ Rcpp::List estimate_theta_1subpop_sample(Rcpp::IntegerMatrix genotypes, bool ret
 
 //' Estimate theta from individuals
 //' 
-//' Estimate theta for one subpopulation given a sample of genotypes.
+//' Estimate theta for one subpopulation given a list of individuals.
 //' 
 //' @inheritParams estimate_theta_1subpop_sample
 //' @param individuals Individuals to get haplotypes for.
@@ -513,5 +513,296 @@ Rcpp::List estimate_theta_1subpop_individuals(Rcpp::ListOf< Rcpp::XPtr<Individua
                                             return_estimation_info);
   return theta;
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+void print_map(std::unordered_map<int, double> x) {
+  for (auto j = x.begin(); j != x.end(); ++j) { 
+    Rcpp::Rcout << "    allele " << j->first << ": " << j->second << std::endl; 
+  } 
+}
+
+void print_container(std::string headline, std::vector< std::unordered_map<int, double> > x) {
+  Rcpp::Rcout << "===========================================\n";
+  Rcpp::Rcout << headline << "\n";
+  Rcpp::Rcout << "===========================================\n";
+  
+  for (auto i = x.begin(); i != x.end(); ++i) { 
+    Rcpp::Rcout << "  subpop " << (i-x.begin()) << std::endl;
+    print_map(*i);
+  }
+}
+
+
+//' Estimate F, theta, and f from subpopulations
+//' 
+//' Estimates F, theta, and f for a number of subpopulations given a list of individuals.
+//' 
+//' Based on Bruce S Weir, Genetic Data Analysis 2, 1996. (GDA2).
+//' 
+//' @param subpops List of subpopulations, each a list of individuals
+//' @param subpops_sizes Size of each subpopulation
+//' 
+//' @return  Estimates of F, theta, and f
+//' 
+//' @export
+// [[Rcpp::export]]
+Rcpp::List estimate_theta_subpops_individuals(Rcpp::List subpops, 
+                                        Rcpp::IntegerVector subpops_sizes) {
+
+  int r = subpops.size();
+  
+  if (r <= 0) {
+    Rcpp::stop("No subpopulations given");
+  }
+  
+  if (subpops_sizes.size() != r) {
+    Rcpp::stop("length(subpops) != length(subpops_sizes)");
+  }
+  
+  if (any(subpops_sizes <= 0).is_true()) {
+    Rcpp::stop("All subpops_sizes must be positive");
+  }
+  
+  /*
+  GDA2, p. 177.
+  */
+  
+  // H_A: Heterozygous probabilities:
+  // H_A[i][l]: Heterozygous probability of allele l in subpopulation i.
+  std::vector< std::unordered_map<int, double> > H_A(r);
+
+  // P_AA: Homozygous probabilities:
+  // P_AA[i][l]: Homozygous probability of allele l in subpopulation i.
+  std::vector< std::unordered_map<int, double> > P_AA(r);
+
+  // p_A: Allele probability:
+  // p_A[i][l]: Allele probability of allele l in subpopulation i.
+  std::vector< std::unordered_map<int, double> > p_A(r);
+
+  std::vector<double> n(r);
+  
+  double n_mean = 0.0;
+  double n_sum = 0.0;
+  double n2_sum = 0.0;
+  
+  ////////////////////////////////////////////////////
[…]
