+++ src/RcppExports.cpp
@@ -140,6 +140,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// estimate_theta_subpops_individuals
+Rcpp::List estimate_theta_subpops_individuals(Rcpp::List subpops, Rcpp::IntegerVector subpops_sizes);
+RcppExport SEXP _malan_estimate_theta_subpops_individuals(SEXP subpopsSEXP, SEXP subpops_sizesSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::List >::type subpops(subpopsSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type subpops_sizes(subpops_sizesSEXP);
+    rcpp_result_gen = Rcpp::wrap(estimate_theta_subpops_individuals(subpops, subpops_sizes));
+    return rcpp_result_gen;
+END_RCPP
+}
 // pedigrees_all_populate_haplotypes
 void pedigrees_all_populate_haplotypes(Rcpp::XPtr< std::vector<Pedigree*> > pedigrees, int loci, Rcpp::NumericVector mutation_rates, bool progress);
 RcppExport SEXP _malan_pedigrees_all_populate_haplotypes(SEXP pedigreesSEXP, SEXP lociSEXP, SEXP mutation_ratesSEXP, SEXP progressSEXP) {
@@ -681,6 +693,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_malan_hash_colisions", (DL_FUNC) &_malan_hash_colisions, 1},
     {"_malan_estimate_theta_1subpop_sample", (DL_FUNC) &_malan_estimate_theta_1subpop_sample, 2},
     {"_malan_estimate_theta_1subpop_individuals", (DL_FUNC) &_malan_estimate_theta_1subpop_individuals, 2},
+    {"_malan_estimate_theta_subpops_individuals", (DL_FUNC) &_malan_estimate_theta_subpops_individuals, 2},
     {"_malan_pedigrees_all_populate_haplotypes", (DL_FUNC) &_malan_pedigrees_all_populate_haplotypes, 4},
     {"_malan_pedigrees_all_populate_haplotypes_custom_founders", (DL_FUNC) &_malan_pedigrees_all_populate_haplotypes_custom_founders, 4},
     {"_malan_pedigrees_all_populate_haplotypes_ladder_bounded", (DL_FUNC) &_malan_pedigrees_all_populate_haplotypes_ladder_bounded, 6},
