+++ tests/testthat/test-autosomal.R
@@ -91,31 +91,31 @@ get_ols_quantities <- function(y) {
 }
 
 y <- t(x)
-ols_res_sample <- get_ols_quantities(y)
-theta_res <- estimate_theta_1subpop_sample(y)
-test_that("estimate_theta_1subpop_sample works", {
+ols_res_genotypes <- get_ols_quantities(y)
+theta_res <- estimate_theta_1subpop_genotypes(y)
+test_that("estimate_theta_1subpop_genotypes works", {
   expect_true(theta_res$error == FALSE)
-  expect_equal(qr.solve(ols_res_sample$x, ols_res_sample$y), 
+  expect_equal(qr.solve(ols_res_genotypes$x, ols_res_genotypes$y), 
                theta_res$estimate)
 })
 
-theta_res_info <- estimate_theta_1subpop_sample(y, return_estimation_info = TRUE)
-test_that("estimate_theta_1subpop_sample return_estimation_info works", {
+theta_res_info <- estimate_theta_1subpop_genotypes(y, return_estimation_info = TRUE)
+test_that("estimate_theta_1subpop_genotypes return_estimation_info works", {
   expect_true(theta_res_info$error == FALSE)
-  expect_equal(qr.solve(ols_res_sample$x, ols_res_sample$y), 
+  expect_equal(qr.solve(ols_res_genotypes$x, ols_res_genotypes$y), 
                theta_res_info$estimate)
-  expect_equal(sort(ols_res_sample$x), sort(theta_res_info$estimation_info$X))
+  expect_equal(sort(ols_res_genotypes$x), sort(theta_res_info$estimation_info$X))
 })
 
 # bootstrap:
 theta_boot <- replicate(100, {
   yboot <- y[sample(nrow(y), replace = TRUE), ]
-  estimate_theta_1subpop_sample(yboot)$estimate
+  estimate_theta_1subpop_genotypes(yboot)$estimate
 })
 # Expanding a bit...
 theta_boot_rng <- c(0.9, 1.1) * range(theta_boot)
 #theta_boot_rng
-test_that("estimate_theta_1subpop_sample boot contains true", {
+test_that("estimate_theta_1subpop_genotypes boot contains true", {
   expect_true(all(theta_boot >= 0 & theta_boot <= 1))
   expect_true(theta >= theta_boot_rng[1L] & theta <= theta_boot_rng[2L])
 })
@@ -143,21 +143,21 @@ test_that("estimate_theta_1subpop_individuals works", {
   expect_equal(qr.solve(ols_res_livepop$x, ols_res_livepop$y),
                estimate_theta_1subpop_individuals(livepop)$estimate, 
                tol = 10*ESTIMATION_TOL_DUE_TO_SAMPLING
-               )
+  )
 })
 
 test_that("livepop: haplotypes/individual same theta", {
-  expect_equal(estimate_theta_1subpop_sample(y_livepop)$estimate,
+  expect_equal(estimate_theta_1subpop_genotypes(y_livepop)$estimate,
                estimate_theta_1subpop_individuals(livepop)$estimate)
 })
 
 
 test_that("geneaology independent sample and population same theta", {
-  expect_equal(estimate_theta_1subpop_sample(y)$estimate,
-               estimate_theta_1subpop_sample(y_livepop)$estimate,
+  expect_equal(estimate_theta_1subpop_genotypes(y)$estimate,
+               estimate_theta_1subpop_genotypes(y_livepop)$estimate,
                tol = ESTIMATION_TOL_DUE_TO_SAMPLING)
[…]
