+++ NAMESPACE
@@ -17,8 +17,10 @@ export(count_brothers)
 export(count_haplotype_occurrences_individuals)
 export(count_haplotype_occurrences_pedigree)
 export(count_uncles)
+export(estimate_theta_1subpop_genotypes)
 export(estimate_theta_1subpop_individuals)
-export(estimate_theta_1subpop_sample)
+export(estimate_theta_subpops_genotypes)
+export(estimate_theta_subpops_individuals)
 export(father_matches)
 export(generate_get_founder_haplotype_db)
 export(generate_get_founder_haplotype_ladder)

[…]
