### 2018-04-27 443c4f3
add more details on XPtr finaliser
