+++ src/malan_types.h
@@ -12,6 +12,13 @@
 #define CHECK_ABORT_EVERY 10000
 
 
+/*
+Rcpp::Xptr<p: pointer, set_delete_finalizer: bool>
+
+set_delete_finalizer: if set to true, a finalizer will be registered for the external pointer. 
+The finalizer is called when the Xptr is garbage collected. 
+The finalizer is merely a call to the delete operator or the pointer so you need to make sure the pointer can be "delete" d this way (has to be a C++ object)
+*/
 //#define RCPP_XPTR_2ND_ARG true // ensures that finaliser is called
 #define RCPP_XPTR_2ND_ARG false // do not call finalisers, I guess we live with some memory leaks for now...!!! FIXME
 
