### 2018-04-27 dde350a
Typos in vignette
