+++ tests/testthat/test-autosomal.R
@@ -209,79 +209,50 @@ test_that("estimate_theta_subpops_individuals = estimate_theta_subpops_genotypes
 })
 
 
-if (FALSE) {
-  library(hierfstat)
-  library(adegenet)
-  data(nancycats, package = "adegenet")
-  fstat(nancycats)
-  
-  library(pegas)
-  as.loci(nancycats)
-  fsttab <- Fst(as.loci(nancycats))
-  fsttab
-  
-  x <- as.loci(nancycats)
-  str(x)
-  
-  
-  grp <- as.integer(x$population)
-  y <- do.call(rbind, strsplit(as.character(x$fca8), "/"))
-  is <- apply(y, 1, function(as) any(is.na(as)))
-  grp <- grp[!is]
-  y <- y[!is, ]
-  str(y)
-  dim(y)
-  lvls <- unique(c(y))
-  #z <- as.data.frame(apply(y, 2, function(as) factor(as, levels = lvls)))
-  z <- apply(y, 2, function(as) match(as, lvls))
-  
-  z_lst <- lapply(split(seq_len(nrow(z)), grp), function(is) z[is, ])
-  z_len <- sapply(z_lst, nrow)
-  estimate_theta_subpops_genotypes(z_lst, z_len)
-  Fst(x)[1, ]
-  
-  fstat(nancycats, fstonly = TRUE)
-  
-  x_nona <- na.omit(x[, c(1, 2, 2)])
-  colnames(x_nona)[3] <- "duploc"
-  x_small <- df2genind(x_nona[, 2:3, F], pop = x_nona[, 1], sep = "/", type = "codom")
-  fstat(x_small, fstonly = TRUE)
-  
-  #https://www2.unil.ch/popgen/softwares/fstat.htm
+############
+# Known answers, GDA2
+two2cols <- function(x) {
+  do.call(rbind, lapply(x, function(as) as.integer(strsplit(as.character(as), "")[[1]])))
 }
-if (FALSE) {
-  ###########
-  res_indv$theta
-  #estimate_theta_subpops_individuals(subpops, c(1, 1))
-  estimate_theta_subpops_individuals(subpops, c(100, 100))$theta
-  estimate_theta_subpops_individuals(subpops, c(1000, 1000))$theta
-  estimate_theta_subpops_individuals(subpops, c(10000, 10000))$theta
-  
-  subpop_haps <- lapply(subpops, get_haplotypes_individuals)
-  hom0 <- lapply(subpop_haps, function(x) mean(apply(x, 1, function(h) all(h == 0L))))
-  hom0
-  het0 <- lapply(subpop_haps, function(x) mean(apply(x, 1, function(h) any(h == 0) && !all(h == 0L))))
-  het0
-  #lapply(subpop_haps, function(x) mean(apply(x, 1, function(h) all(h == 1L))))
-  #lapply(subpop_haps, function(x) mean(apply(x, 1, function(h) all(h == 2L))))
-  
-  subpops_haps <- lapply(split(seq_len(nrow(y_livepop)), grps), function(is) y_livepop[is, ])
-  subpops_haps_sizes <- sapply(subpops_haps, nrow)
-  
-  estimate_theta_subpops_genotypes(subpops_haps, subpops_haps_sizes)
-  ###########
-  
-  grp1 <- lapply(get_pids_in_pedigree(peds[[1]]), function(pid) get_individual(sim_res_fixed$population, pid))
[…]
