### 2018-04-27 2e8c198
subpop theta by pid
