+++ NAMESPACE
@@ -21,6 +21,7 @@ export(estimate_theta_1subpop_genotypes)
 export(estimate_theta_1subpop_individuals)
 export(estimate_theta_subpops_genotypes)
 export(estimate_theta_subpops_individuals)
+export(estimate_theta_subpops_pids)
 export(father_matches)
 export(generate_get_founder_haplotype_db)
 export(generate_get_founder_haplotype_ladder)
