### 2018-04-29 c4e9d1e
split by pid and haplotype hash
