+++ man/split_by_haplotypes.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/RcppExports.R
+\name{split_by_haplotypes}
+\alias{split_by_haplotypes}
+\title{Split pids by haplotype}
+\usage{
+split_by_haplotypes(population, pids)
+}
+\arguments{
+\item{population}{Population obtained from simulation}
+
+\item{pids}{Vector of individual pids}
+}
+\value{
+List of integer vector, element i is an IntegerVector
+with all pids from \code{pids} with the same haplotype
+}
+\description{
+Individuals with the same haplotype will be in the same group
+and individuals with different haplotypes will be in different groups.
+}
