+++ man/haplotypes_to_hashes.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/RcppExports.R
+\name{haplotypes_to_hashes}
+\alias{haplotypes_to_hashes}
+\title{Convert haplotypes to hashes (integers)}
+\usage{
+haplotypes_to_hashes(population, pids)
+}
+\arguments{
+\item{population}{Population obtained from simulation}
+
+\item{pids}{Vector of individual pids}
+}
+\value{
+Integer vector with haplotype hashes
+}
+\description{
+Individuals with the same haplotype will have the same hash (integer)
+and individuals with different haplotypes will have different hashes (integers).
+}
+\details{
+This can be useful if for example using haplotypes to define groups
+and the haplotype itself is not of interest.
+}

[…]
