+++ tests/testthat/test-simulate-population.R
@@ -3,7 +3,7 @@ context("Simulate population")
 set.seed(1)
 sim_res_fixed <- sample_geneology(population_size = 1e3, 
                                   generations = 20, 
-                                  extra_generations_full = 2,
+                                  generations_full = 3,
                                   individuals_generations_return = 2, # default value
                                   progress = FALSE)
 
@@ -20,7 +20,7 @@ sim_res_growth <- sample_geneology_varying_size(population_sizes = rep(1e3, 20),
                                                 enable_gamma_variance_extension = TRUE,
                                                 gamma_parameter_shape = 5,
                                                 gamma_parameter_scale = 1/5,
-                                                extra_generations_full = 2,
+                                                generations_full = 3,
                                                 individuals_generations_return = 2, # default value
                                                 progress = FALSE)
 
