+++ inst/doc/introduction.R
@@ -90,7 +90,7 @@ set.seed(1)
 sim_res <- sample_geneology(population_size = 1e3, 
                             generations = 200, 
                             generations_full = 3,
-                            individuals_generations_return = 2, # default value
+                            generations_return = 3, # default value
                             progress = FALSE)
 
 ## ------------------------------------------------------------------------
