+++ man/sample_geneology.Rd
@@ -33,7 +33,7 @@ sample_geneology(population_size, generations, generations_full = 1L,
 \item{verbose_result}{Verbose result.}
 }
 \value{
-A list with the following entries:
+A malan_simulation / list with the following entries:
 \itemize{
 \item \code{population}. An external pointer to the population.
 \item \code{generations}. Generations actually simulated, mostly useful when parameter \code{generations = -1}.
@@ -41,7 +41,7 @@ A list with the following entries:
 \item \code{growth_type}. Growth type model.
 \item \code{sdo_type}. Standard deviation in a man's number of male offspring. StandardWF or GammaVariation depending on \code{enable_gamma_variance_extension}.
 \item \code{end_generation_individuals}. Pointers to individuals in end generation.
-\item \code{individuals_generations}. Pointers to individuals in end generation in addition to the previous \code{individuals_generations_return}.
+\item \code{individuals_generations}. Pointers to individuals in last \code{generations_return} generation (if \code{generations_return = 3}, then individuals in the last three generations are returned).
 }
 If \code{verbose_result} is true, then these additional components are also returned:
 \itemize{
