+++ src/api_utility_autosomal.cpp
@@ -549,10 +549,6 @@ void print_container(std::string headline, std::vector< std::unordered_map<int, 
 
 
 /*
-// H_A: Heterozygous probabilities:
-// H_A[i][l]: Heterozygous probability of allele l in subpopulation i.
-std::vector< std::unordered_map<int, double> > H_A(r);
-
 // P_AA: Homozygous probabilities:
 // P_AA[i][l]: Homozygous probability of allele l in subpopulation i.
 std::vector< std::unordered_map<int, double> > P_AA(r);
@@ -561,12 +557,11 @@ std::vector< std::unordered_map<int, double> > P_AA(r);
 // p_A[i][l]: Allele probability of allele l in subpopulation i.
 */
 Rcpp::List estimate_theta_subpops_weighted_engine(
-    std::vector< std::unordered_map<int, double> > H_A,
     std::vector< std::unordered_map<int, double> > P_AA,
     std::vector< std::unordered_map<int, double> > p_A,
     std::vector<double> n) {
   
-  int r = H_A.size();
+  int r = P_AA.size();
   
   if (r <= 0) {
     Rcpp::stop("r <= 0");
@@ -596,41 +591,6 @@ Rcpp::List estimate_theta_subpops_weighted_engine(
     n2_sum += n_i * n_i;
   }
 
-  ////////////////////////////////////////////////////
-  // Calculating helper variables
-  ////////////////////////////////////////////////////
-  
-  //********************************
-  // GDA2, p. 178, H_A. tilde
-  //********************************
-  std::unordered_map<int, double> mean_H_A;
-  for (int i = 0; i < r; ++i) {
-    for (auto ele = H_A[i].begin(); ele != H_A[i].end(); ++ele) {
-      int allele = ele->first;      
-      double HAi = ele->second;
-      
-      mean_H_A[allele] += (n[i] * HAi) / n_sum;
-    } 
-  }  
-  //Rcpp::Rcout << "mean_H_A\n";
-  //print_map(mean_H_A);
-  
-  /*
-  // Gives the same as mean_H_A
-  std::unordered_map<int, double> mean_H_A_type2;
-  for (int i = 0; i < r; ++i) {
-  for (auto ele = H_A[i].begin(); ele != H_A[i].end(); ++ele) {
-  int allele = ele->first;      
-  double HAi = ele->second;
-  alleles.insert(allele);
-  
-  mean_H_A_type2[allele] += (2.0 * n[i] * (p_A[i][allele] - P_AA[i][allele]) ) / n_sum;
-  } 
-  }  
-  Rcpp::Rcout << "mean_H_A_type2\n";
-  print_map(mean_H_A_type2);
-  */
-  
   // So have a common container with alleles to iterate over later
   std::unordered_set<int> alleles;
   
@@ -644,9 +604,10 @@ Rcpp::List estimate_theta_subpops_weighted_engine(
       mean_pA[ ele->first ] += (n[i] * ele->second) / n_sum;
     } 
   }  
-  //Rcpp::Rcout << "mean_pA\n";
-  //print_map(mean_pA);
-  
+  /*
+  Rcpp::Rcout << "mean_pA\n";
+  print_map(mean_pA);
+  */
   //********************************
   // GDA2, p. 173, s^2
   //********************************
@@ -658,21 +619,55 @@ Rcpp::List estimate_theta_subpops_weighted_engine(
       s2_A[ ele->first ] += (n[i] * d * d) / ((r-1.0) * n_mean);
     } 
[…]
