+++ man/estimate_theta_subpops_pids.Rd
@@ -14,7 +14,7 @@ estimate_theta_subpops_pids(population, subpops, subpops_sizes)
 \item{subpops_sizes}{Size of each subpopulation}
 }
 \value{
-Estimates of F, theta, and f
+Estimates of F, theta, and f as well as sums of squares S1, S2 and S3
 }
 \description{
 Estimates F, theta, and f for a number of subpopulations given a list of pids (individual ids).
