### 2018-05-01 2f9583c
autosomal: additional tests, doc and clean up
