+++ src/api_utility_autosomal.cpp
@@ -579,6 +579,8 @@ Rcpp::List estimate_theta_subpops_weighted_engine(
     Rcpp::stop("n.size() != r");
   }
   
+  double r_dbl = (double)r;
+  
   double n_mean = 0.0;
   double n_sum = 0.0;
   double n2_sum = 0.0;
@@ -586,7 +588,7 @@ Rcpp::List estimate_theta_subpops_weighted_engine(
   for (int i = 0; i < r; ++i) {
     // Different from subpop.size()!
     double n_i = n[i];
-    n_mean += n_i / (double)r;
+    n_mean += n_i / r_dbl;
     n_sum += n_i;
     n2_sum += n_i * n_i;
   }
@@ -594,35 +596,37 @@ Rcpp::List estimate_theta_subpops_weighted_engine(
   // So have a common container with alleles to iterate over later
   std::unordered_set<int> alleles;
   
+  // First, be sure that all alleles are known:
+  for (int i = 0; i < r; ++i) {
+    for (auto ele = p_A[i].begin(); ele != p_A[i].end(); ++ele) { 
+      alleles.insert(ele->first);
+    }
+  }
+  
   //********************************
   // GDA2, p. 168, p_A. tilde
   //********************************
   std::unordered_map<int, double> mean_pA;
   for (int i = 0; i < r; ++i) {
-    for (auto ele = p_A[i].begin(); ele != p_A[i].end(); ++ele) { 
-      alleles.insert(ele->first);
-      mean_pA[ ele->first ] += (n[i] * ele->second) / n_sum;
+    for (auto ele = alleles.begin(); ele != alleles.end(); ++ele) {
+      int allele = *ele;
+      double p = p_A[i][allele]; // 0 if not exists
+      mean_pA[allele] += (n[i] * p) / n_sum;
     } 
   }  
-  /*
-  Rcpp::Rcout << "mean_pA\n";
-  print_map(mean_pA);
-  */
+
   //********************************
   // GDA2, p. 173, s^2
   //********************************
   std::unordered_map<int, double> s2_A;
   for (int i = 0; i < r; ++i) {
-    for (auto ele = p_A[i].begin(); ele != p_A[i].end(); ++ele) {
-      alleles.insert(ele->first);
-      double d = ele->second - mean_pA[ele->first];
-      s2_A[ ele->first ] += (n[i] * d * d) / ((r-1.0) * n_mean);
+    for (auto ele = alleles.begin(); ele != alleles.end(); ++ele) {
+      int allele = *ele;
+      double p = p_A[i][allele]; // 0 if not exists
+      double d = p - mean_pA[allele];
+      s2_A[allele] += (n[i] * d * d) / ((r_dbl - 1.0) * n_mean);
     } 
   }  
-  /*
-  Rcpp::Rcout << "s2_A\n";
-  print_map(s2_A);
-  */
 
   
   ////////////////////////////////////////////////////
@@ -635,32 +639,23 @@ Rcpp::List estimate_theta_subpops_weighted_engine(
   std::unordered_map<int, double> mean_H_A;
   for (auto ele = alleles.begin(); ele != alleles.end(); ++ele) {
     int allele = *ele;
+    
     for (int i = 0; i < r; ++i) {
       double pA = p_A[i][allele];
       double PAA = P_AA[i][allele];
[…]
