### 2018-05-01 #684 @ajank
@mikldk, you may wish to fix the typo in the description at https://github.com/mikldk/malan as well.
