### 2018-05-01 96331cb
Tests for allele wise theta/F values
