### 2018-05-01 d558c1f
Removing compiler warning about anonymous namespace
