### 2018-05-01 c73b6a0
DOI's in references where applicable
