+++ paper.bib
@@ -5,7 +5,9 @@
 	Title = {{How convincing is a matching Y-chromosome profile?}},
 	Volume = {13},
 	Number = {11},
-	Year = {2017}}
+	Year = {2017},
+	doi = {10.1371/journal.pgen.1007028}
+}
 
 @book{DJBwoe2,
 	Author = {Steele, CD and Balding, DJ},
@@ -14,7 +16,9 @@
 	Edition = {2nd ed.},
 	Publisher = {Wiley},
 	Title = {Weight of evidence for forensic {DNA} profiles},
-	Year = 2015}
+	Year = 2015,	
+	doi = {10.1002/9780470867693}
+}
 
 @manual{R,
 	Address = {Vienna, Austria},
@@ -23,7 +27,7 @@
 	Title = {R: A Language and Environment for Statistical Computing},
 	Url = {https://www.R-project.org/},
 	Year = {2018},
-	Bdsk-Url-1 = {https://www.R-project.org/}}
+}
 
 @Article{Rcpp,
   title = {{Extending \textit{R} with \textit{C++}: A Brief Introduction to \textit{Rcpp}}},
@@ -35,16 +39,17 @@
   pages = {e3188v1},
   issn = {2167-9843},
   url = {https://doi.org/10.7287/peerj.preprints.3188v1},
-  doi = {10.7287/peerj.preprints.3188v1},
+  doi = {10.7287/peerj.preprints.3188v1}
 }
 
 @article{Butler05,
 	Author = {Butler, JM and Decker, AE and Kline, MC and Vallone, PM},
-	Date-Added = {2017-01-19 09:38:22 +0000},
-	Date-Modified = {2017-02-04 01:34:09 +0000},
-	Journal = {J Forensic Sci},
+	Journal = {Journal of Forensic Sciences},
 	Number = {4},
 	Pages = {853-9},
 	Title = {Chromosomal duplications along the {Y} chromosome and their potential impact on {Y-STR} interpretation},
 	Volume = {50},
-	Year = {2005}}
+	Year = {2005},	
+	doi = {10.1520/JFS2004481}
+}
+
