### 2018-05-01 2358231
Insert Rcpp next to reference to it
