### 2018-05-01 908bea6
Remove abbreviated affiliations
