### 2018-05-01 #684 @mikldk
@ajank , also thanks for your suggestions, they er now all addressed.
