### 2018-05-01 #684 @mikldk
@ajank , Sorry, but I don't understand what you mean. Do you mean that it would be easier for you that I had waited until you said you were finished? In that case, I am sorry. I thought it was easier to do the changes and notify you about them instantly. Sorry about that.
