### 2018-05-01 #684 @ajank
@mikldk, maybe I should have added a smiley here ;) being new to JOSS, it was an experience on its own for me! No need to worry :)
