+++ .travis.yml
@@ -27,7 +27,7 @@ after_script: R -q -e 'tic::after_script()'
 # Header
 language: r
 sudo: true
-dist: trusty
+#dist: trusty
 cache: packages
 
 matrix:
@@ -37,6 +37,11 @@ matrix:
     env:
     - BUILD_PKGDOWN: true
   - r: devel
+  - os: linux
+    dist: trusty
+    sudo: required
+  - os: osx
+    osx_image: xcode8.3
 
 #env
 env:
