### 2018-05-02 7cd2f01
adding macosx to travis
