+++ .travis.yml
@@ -27,21 +27,30 @@ after_script: R -q -e 'tic::after_script()'
 # Header
 language: r
 sudo: true
-dist: trusty
 cache: packages
 
 matrix:
   include:
-  - r: oldrel
-  - r: release
-    env:
-    - BUILD_PKGDOWN: true
-  - r: devel
-  - os: linux
-    dist: trusty
-    sudo: required
-  - os: osx
-    osx_image: xcode8.3
+    - os: linux
+      dist: trusty
+      r: oldrel
+    - os: linux
+      dist: trusty
+      r: release
+      env: 
+        - BUILD_PKGDOWN: true
+    - os: linux
+      dist: trusty
+      r: devel
+    - os: osx
+      osx_image: xcode8.3
+      r: oldrel
+    - os: osx
+      osx_image: xcode8.3
+      r: release
+    - os: osx
+      osx_image: xcode8.3
+      r: devel
 
 #env
 env:
