### 2018-05-04 #684 @afrubin
@mikldk Thanks for your prompt response to the build-related GitHub issues. There are a few JOSS-specific things I'd like to bring up here, as they are relate to the checklist rather than the code itself.

- The package version matches, but there is no GitHub release with that version number.
- While the installation instructions are clear and dependencies are handled by R, dependencies in the `DESCRIPTION` should be listed in the documentation.
- I didn't see a Zenodo or similar link stable to the release version of the software. This should be created and provided in the paper.
