+++ README.md
@@ -49,7 +49,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 ## License
 
-License: GPL (>= 2).
+License: GPL-2.
 
 ## Build status
 
