+++ README.md
@@ -31,6 +31,18 @@ To read a vignette, type:
 ```r
 vignette("introduction", package = "malan")
 ```
+### Running tests
+
+Note that to also install the tests, you need to install the package as follows:
+
+``` r
+# install.packages("remotes")
+remotes::install_github("mikldk/malan", INSTALL_opts="--install-tests")
+
+library(malan)
+library(testthat)
+test_package('malan')
+```
 
 ## Contribute, issues, and support
 
