### 2018-05-04 #684 @mikldk
@afrubin Thank you very much.

* Github release: Sorry. Now, there is: <https://github.com/mikldk/malan/releases/tag/0.0.3>
* Dependencies: Added in mikl/malan@d92e8f35ca1f72b972c988ed49a335af49b60869.
* Zenodo: Sorry, now, there is:  https://zenodo.org/record/1240896 / DOI: 10.5281/zenodo.1240896 <http://doi.org/10.5281/zenodo.1240896> (DOI doesn't work, yet)
