### 2018-05-04 #684 @mikldk
@brainstorm For some reason it seems like I can edit the reviewers checklists (I haven't dared trying, but the `checkbox`es are not disabled).

@afrubin Stupid question: So now that the version 0.0.3 is locked and have a DOI number. What happens if there is now any blocking changes that I need to do in this review? Do the changes, delete the Github release and create that same 0.0.3 again? And when the paper is hopefully accepted at some point, I should increment the version number for future changes, to e.g. 0.0.4 or whatever, right? Or how should this version numbering / DOI / publishing workflow be done?
