### 2018-05-04 #684 @mikldk
@afrubin Thanks. But if I now add a badge then that change will not be in the release. But maybe it doesn't matter dor such small changes?
