### 2018-05-04 #684 @arfon
> @afrubin Thanks. But if I now add a badge then that change will not be in the release. But maybe it doesn't matter dor such small changes?

I think it's OK for the release number we're quoting here not to have the badge in the README.
