### 2018-05-06 5931866
Bumping version to 1.0.0
