+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: malan
 Type: Package
 Title: MAle Lineage ANalysis
-Version: 0.0.3
-Date: 2018-04-24
+Version: 1.0.0
+Date: 2018-05-06
 Authors@R: c(
     person(given = "Mikkel Meyer", family = "Andersen", 
         email = "mikl@math.aau.dk", role = c("aut", "cre"))
