### 2018-05-06 bce5182
Updateting zenodo badge link to release v1.0.0
