+++ README.md
@@ -84,5 +84,5 @@ Travis CI:
 
 Zenodo: 
 
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1240896.svg)](https://doi.org/10.5281/zenodo.1240896)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1241769.svg)](https://doi.org/10.5281/zenodo.1241769)
 
