### 2018-05-06 #684 @mikldk
@afrubin I hope that I have now addressed all your remarks. I have updated the software to version 1.0.0 including `DESCRIPTION` and `codemeta.json` as well as made a new release at Github and updated `README` with the new DOI from Zenodo.

@afrubin and @ajank : Thank you both for helping me improving the software. Should you have any more comments, please let me know either here or by opening a new issue in `malan`'s repository.
