### 2018-05-06 #684 @afrubin
@mikldk All of the major outstanding issues have been addressed and I am happy for the paper to proceed to the next stages. Thank your quick feedback and assistance with the Mac OS build issues in particular. The documentation has improved quite a lot during the review, and I hope that this will help your future users.

There are two very minor points that I think should be addressed before publication:
- The version of the software being published is now 1.0.0 instead of 0.0.3
- The Zenodo DOI for the 1.0.0 release should be included in the paper.md (@arfon is this done automatically when the PDF is built?)
