### 2018-05-06 #684 @mikldk
@afrubin You haven't checked the References part. Am I missing something there, you think? Or is it related to the versioning?
