### 2018-05-06 #684 @afrubin
> Or is it related to the versioning?

@mikldk I hadn't checked this because I didn't see the DOI for the package in the paper, but having looked at some published JOSS papers, it does appear that it's added automatically to the PDF. I've now checked this box.
