### 2018-05-06 #684 @arfon
Thanks folks. I've gone ahead and marked the version as `v1.0.0` at the top of the issue.

@brainstorm - I think we're good to accept here?
