### 2018-05-07 #684 @brainstorm
Thanks everyone! Closing the issue, everything looks fine to me, @arfon, do your DOI publishing magic ;)
