### 2018-05-07 #684 @mikldk
@arfon : I hereby remind you about #679: the footer in the pdf article proof contains the Markdown markup. Can you remove this?
