### 2018-05-07 #684 @arfon
> Thanks everyone! Closing the issue, everything looks fine to me, @arfon, do your DOI publishing magic ;)

Thanks @brainstorm - please don't close these until I've wrapped up the last few steps. :-)
